Provide `prepare` front-end functionality
Workflows and blocks are looked for in `./.ripe` wherever invoked, directories
specified in the `$RIPELIB` environment variable.

`.ripe/tasks/count.sh` is a block template.

`.ripe/workflows/star.rb` is a workflow template that contains two
variables in global space

  * `$vars` are workflow-wide configurations (i.e. resource allocation,
    checkpoint restart mode)
  * `$callback` takes a sample name and should return a `WorkingBlock`

It is applied to a sample set by calling

  ripe prepare -w star -m patch SAMPLE1 SAMPLE2 [...]

To use a block/task in a workflow, use the `Ripe::Task` function.

Lastly, to avoid confusion, worker data has been moved to `./ripe/workers`
instead of `.ripe`.