Produce per-task job script
Prior to this change, only a per-worker job script is outputted,
containing the entire job to be run on a node, with a path defined in
`Worker#sh`.  Now, per-task job scripts containing only the portion of
the script defining their task is outputted, with a path defined by
`Task#sh`.

This closes #1.